verify_domains: compute symmetric diff between csr and command line domains
Computing the set difference only catches when the csr is missing a domain
that is listed on the command line.  It fails to catch when the csr has
one or more *extra* domains than appear on the command line.  This results
in one or more extra hosts being listed in the csr than are on the command
line.

This is bad since it could either issue a cert with more domains than you
asked for if the extra domains pass their challenge, or it can also fail
if the extra (unwanted) domains fail their challenge.

Changed to use symmetric_difference which will catch any domains that are
in either set but not in both.