generate: loop over generation of csr and crt in case verify_csr/crt fails
When verification of the csr or crt fails, the user is given the option
of removing the related files and continuing.  If the user chooses to
remove the files, the very next step in the process fails since it
depends on the files generated by the previous step.

This change will loop to regenerate the files if the user chooses to
delete them.